Random code snippets using predefined code templates
This code generates a random code snippet using a list of pre-defined code templates. The templates are strings that contain placeholders, indicated by curly braces ({}), that will be replaced with random values to create a complete code snippet.

The datetime and random modules are used to seed the random number generator, ensuring that the same code snippet is not generated each time the program is run. The today variable is created by formatting the current date and time as a string with the format '%Y%m%d'. This string is then used as the seed for the random module.

The program then selects a random template from the templates list using the random.choice() function. If the template contains any placeholders, indicated by the presence of '{}', the program generates a list of random placeholder